+++ flang/lib/Lower/ConvertCall.cpp
@@ -1135,7 +1135,8 @@ genUserCall(PreparedActualArguments &loweredActuals,
         // callee side, and it is illegal to use NULL without a MOLD if any
         // dummy length parameters are assumed.
         mlir::Type boxTy = fir::dyn_cast_ptrEleTy(argTy);
-        assert(boxTy && boxTy.isa<fir::BoxType>() && "must be a fir.box type");
+        assert(boxTy && boxTy.isa<fir::BaseBoxType>() &&
+               "must be a fir.box type");
         mlir::Value boxStorage = builder.createTemporary(loc, boxTy);
         mlir::Value nullBox = fir::factory::createUnallocatedBox(
             builder, loc, boxTy, /*nonDeferredParams=*/{});